Move channel info commands to new application
