+++ src/Prima.Stable/Modules/ExtraModule.cs
@@ -108,29 +108,5 @@ namespace Prima.Stable.Modules
             public int Total { get; set; }
             public string WorldName { get; set; }
         }
-
-        [Command("setdescription")]
-        [RequireUserPermission(GuildPermission.BanMembers)]
-        public async Task SetDescriptionAsync([Remainder] string description)
-        {
-            await Db.DeleteChannelDescription(Context.Channel.Id);
-            await Db.AddChannelDescription(Context.Channel.Id, description);
-            await ReplyAsync($"{Context.User.Mention}, the help message has been updated!");
-        }
-
-        [Command("whatisthis")]
-        [Description("Explains what the channel you use it in is for, if such information is available.")]
-        public Task WhatIsThisAsync()
-        {
-            var cd = Db.ChannelDescriptions.FirstOrDefault(cd => cd.ChannelId == Context.Channel.Id);
-            if (cd == null) return Task.CompletedTask;
-            var embed = new EmbedBuilder()
-                .WithTitle($"#{Context.Channel.Name}")
-                .WithColor(new Color(0x00, 0x80, 0xFF))
-                .WithThumbnailUrl("http://www.newdesignfile.com/postpic/2016/05/windows-8-help-icon_398417.png")
-                .WithDescription(cd.Description)
-                .Build();
-            return ReplyAsync(embed: embed);
-        }
     }
 }
